% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Filter by \%notin\% similar to \%in\%}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector: the values to be matched}

\item{y}{vector: the values to be matched against}
}
\value{
A vector of the same length as x.
}
\description{
\code{\%notin\%} returns a logical vector indicating \code{TRUE} if there is
not a match for its left operand
}
\examples{
1:10 \%notin\% c(1,3,5,9)

}
