% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_rename.R
\name{lapplyWithRename}
\alias{lapplyWithRename}
\title{\code{lapply} and rename list elements}
\usage{
lapplyWithRename(...)
}
\arguments{
\item{...}{Arguments passed to lapply}
}
\value{
\code{list} with elements named bast on the elements of
     the first argument to \code{lapplyWithRename}
}
\description{
\code{lapplyWithRename} works the same as \code{lapply} except it renames
the elements of the output with the names of the elements of the input
}
\examples{
f <- function(x) structure(x, .Names = paste("Name", x))
L <- list(a="A", b="B")
lapplyWithRename(L, f) 

}
