% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtry_finder.R
\name{mtry_finder}
\alias{mtry_finder}
\title{Calculate mtry for randomForest}
\usage{
mtry_finder(df)
}
\arguments{
\item{df}{\code{data.frame}}
}
\value{
Single numeric value for mtry
}
\description{
Calculates mtry, the number of variables to try at each split, from
a \code{data.frame} for \code{randomForest}.
}
\examples{
data(iris)
mtry <- mtry_finder(iris)

}
