% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_str.R
\name{reverse_str}
\alias{reverse_str}
\title{Reverse a string}
\usage{
reverse_str(string)
}
\arguments{
\item{string}{vector of strings to be reversed}
}
\value{
Vector of strings the same length as \code{string}
}
\description{
\code{reverse_str} takes a vector of strings and returns those strings backwards
}
\examples{
x <- 'abcd'
reverse_str(x)

}
