% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculate root mean squared error}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{observed}{vector of observed numeric response values}

\item{predicted}{vector of numeric values output from predict() or similar}
}
\value{
Single numeric value of the root mean squared error
}
\description{
Calculates root mean squared error from observed and predicted values from a model
}
\examples{
require(stats)
x <- rnorm(10, 1, 1)
y <- rnorm(10, x * 0.5, 1)
fit <- lm(y ~ x)
pred_y <- predict(fit)
rmse(y, pred_y)

}
