% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{whether or not to remove NAs, default is FALSE.}
}
\value{
Single numeric value of the standard error
}
\description{
Calculates standard error based on the standard deviation and number of elements
}
\examples{
x <- rnorm(10, 1, 1)
se(x)

}
